labelling_pipeline/ocr_api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
boto3==1.35.90
botocore==1.35.90
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.0
craft-text-detector==0.4.3
cycler==0.12.1
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gdown==5.2.0
huggingface-hub==0.27.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.5.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.21.6
opencv-python==4.5.4.60
pandas==2.2.3
pillow==11.0.0
py-cpuinfo==9.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==307
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
safetensors==0.4.5
scipy==1.13.1
seaborn==0.13.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==1.12.1
torchaudio==2.5.1
torchvision==0.13.1
tqdm==4.67.1
transformers==4.47.1
tzdata==2024.2
ultralytics==8.3.57
ultralytics-thop==2.0.13
urllib3==1.26.20